Refactor options
