+++ src/SpiffyAuthorize/Factory/GuardRouteFactory.php
@@ -16,8 +16,8 @@ class GuardRouteFactory implements FactoryInterface
      */
     public function createService(ServiceLocatorInterface $serviceLocator)
     {
-        /** @var \SpiffyAuthorize\ModuleOptions $options */
-        $options = $serviceLocator->get('SpiffyAuthorize\ModuleOptions');
+        /** @var \SpiffyAuthorize\Options\ModuleOptions $options */
+        $options = $serviceLocator->get('SpiffyAuthorize\Options\ModuleOptions');
         $guard   = new RouteGuard();
         $guard->setAuthorizeService($serviceLocator->get($options->getAuthorizeService()));
