+++ src/SpiffyAuthorize/Factory/ProviderRoleFactory.php
@@ -2,7 +2,7 @@
 
 namespace SpiffyAuthorize\Factory;
 
-use SpiffyAuthorize\ModuleOptions;
+use SpiffyAuthorize\Options\ModuleOptions;
 
 class ProviderRoleFactory extends AbstractInstanceFactory
 {
