+++ src/SpiffyAuthorize/Factory/RbacServiceFactory.php
@@ -17,8 +17,8 @@ class RbacServiceFactory implements FactoryInterface
      */
     public function createService(ServiceLocatorInterface $serviceLocator)
     {
-        /** @var \SpiffyAuthorize\ModuleOptions $options */
-        $options   = $serviceLocator->get('SpiffyAuthorize\ModuleOptions');
+        /** @var \SpiffyAuthorize\Options\ModuleOptions $options */
+        $options   = $serviceLocator->get('SpiffyAuthorize\Options\ModuleOptions');
         $perms     = $serviceLocator->get('SpiffyAuthorize\PermissionProviders');
         $roles     = $serviceLocator->get('SpiffyAuthorize\RoleProviders');
         $providers = array_merge($perms, $roles);
