+++ src/SpiffyAuthorize/Factory/ViewStrategyUnauthorizedFactory.php
@@ -16,8 +16,8 @@ class ViewStrategyUnauthorizedFactory implements FactoryInterface
      */
     public function createService(ServiceLocatorInterface $serviceLocator)
     {
-        /** @var \SpiffyAuthorize\ModuleOptions $options */
-        $options  = $serviceLocator->get('SpiffyAuthorize\ModuleOptions');
+        /** @var \SpiffyAuthorize\Options\ModuleOptions $options */
+        $options  = $serviceLocator->get('SpiffyAuthorize\Options\ModuleOptions');
         $strategy = new UnauthorizedStrategy();
         $strategy->setTemplate($options->getViewTemplate());
