+++ tests/SpiffyAuthorizeTest/Factory/GuardRouteParamsFactoryTest.php
@@ -3,10 +3,9 @@
 namespace SpiffyAuthorizeTest\Factory;
 
 use Mockery as m;
-use SpiffyAuthorize\ModuleOptions;
+use SpiffyAuthorize\Options\ModuleOptions;
 use SpiffyAuthorize\Factory\GuardRouteParamsFactory;
 use SpiffyAuthorizeTest\Asset\AuthorizeService;
-use Zend\ServiceManager\ServiceManager;
 
 class GuardRouteParamsFactoryTest extends \PHPUnit_Framework_TestCase
 {
@@ -17,7 +16,7 @@ class GuardRouteParamsFactoryTest extends \PHPUnit_Framework_TestCase
 
         $sm = m::mock('Zend\ServiceManager\ServiceManager');
         $sm->shouldReceive('get')->with('AuthorizeService')->andReturn(new AuthorizeService());
-        $sm->shouldReceive('get')->with('SpiffyAuthorize\ModuleOptions')->andReturn($options);
+        $sm->shouldReceive('get')->with('SpiffyAuthorize\Options\ModuleOptions')->andReturn($options);
 
         $factory  = new GuardRouteParamsFactory();
         $instance = $factory->createService($sm);
