+++ tests/SpiffyAuthorizeTest/Factory/ViewStrategyUnauthorizedFactoryTest.php
@@ -3,9 +3,8 @@
 namespace SpiffyAuthorizeTest\Factory;
 
 use Mockery as m;
-use SpiffyAuthorize\ModuleOptions;
+use SpiffyAuthorize\Options\ModuleOptions;
 use SpiffyAuthorize\Factory\ViewStrategyUnauthorizedFactory;
-use Zend\ServiceManager\ServiceManager;
 
 class ViewStrategyUnauthorizedFactoryTest extends \PHPUnit_Framework_TestCase
 {
@@ -14,7 +13,7 @@ class ViewStrategyUnauthorizedFactoryTest extends \PHPUnit_Framework_TestCase
         $options = new ModuleOptions();
 
         $sm = m::mock('Zend\ServiceManager\ServiceManager');
-        $sm->shouldReceive('get')->with('SpiffyAuthorize\ModuleOptions')->andReturn($options);
+        $sm->shouldReceive('get')->with('SpiffyAuthorize\Options\ModuleOptions')->andReturn($options);
 
         $factory  = new ViewStrategyUnauthorizedFactory();
         $instance = $factory->createService($sm);
